Refactored js added new feature
Added new feature to indicate if row, col, or square contains the same
digit more than once.  Found that is Row, Col, and square methods were
repetitive.  Refactored the row_ id to a class so I didn't have to use
the different methods when trying to find the row of an element.